+++ gdsfactory/samples/03_move.py
@@ -2,8 +2,8 @@
 
 # Manipulating geometry 1 - Basic movement and rotation
 
-There are several actions we can take to move and rotate the geometry.  These
-actions include movement, rotation, and reflection.
+There are several actions we can take to move and rotate the geometry.
+These actions include movement, rotation, and reflection.
 
 """