+++ gdsfactory/samples/06_remapping_layers.py
@@ -1,4 +1,4 @@
-""" You can remap layers."""
+"""You can remap layers."""
 import gdsfactory as gf
 from gdsfactory.component import Component