+++ gdsfactory/samples/08_group.py
@@ -1,4 +1,7 @@
-"""Group references. Distribute them ...
+"""Group references. Distribute them as you want
+
+Lets distribute different references in the X direction
+
 """
 
 if __name__ == "__main__":