+++ gdsfactory/samples/21_add_fiber_array.py
@@ -1,5 +1,4 @@
-"""Connecting a component with I/O.
-"""
+"""You can a component with a fiber array."""
 
 import gdsfactory as gf
 from gdsfactory.component import Component